src/com/none/algorithm/sort/Selection.java
package com.none.algorithm.sort;

import static com.none.tool.StdOut.print;
import static com.none.tool.StdOut.println;

/**
 * Created by Nathon on 2017/8/17.
 */
public class Selection {
    public static void sort(Comparable[] arr) {
        int len = arr.length;
        for(int i = 0; i < len; i++) {
            int min = i;
            for(int j = i; j < len; j++) {
                if(less(arr[j], arr[min])) {
                    min = j;
                }
            }
            exch(arr, min, i);
        }
    }

    private static boolean less(Comparable v, Comparable w) {
        return v.compareTo(w) < 0;
    }

    private static void exch(Comparable[] arr, int i, int j) {
        Comparable tmp = arr[i];
        arr[i] = arr[j];
        arr[j] = tmp;
    }

    public static void show(Comparable[] arr) {
        for(Comparable a: arr) {
            print(a + " ");
        }
        println();
    }

    public static boolean isSorted(Comparable[] arr) {
        int len = arr.length;
        for(int i = 1; i < len; i++) {
            if(less(arr[i], arr[i-1])) {
                return false;
            }
        }

        return true;
    }
}
